+++ example/lib/user_profile.widget.dart
@@ -5,25 +5,14 @@
 // StateWidgetGenerator
 // **************************************************************************
 
-// ignore_for_file: type=lint, unused_element, unused_element_parameter, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark, invalid_use_of_visible_for_testing_member, invalid_use_of_protected_member, unnecessary_import, unused_import
-
+// ignore_for_file: type=lint, unused_element, unused_element_parameter, unnecessary_import, unused_import, invalid_use_of_protected_member, invalid_use_of_visible_for_testing_member
 // coverage:ignore-file
-
-import 'package:autoverpod/autoverpod.dart';
-
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-import 'package:riverpod_annotation/riverpod_annotation.dart';
-
 import 'package:flutter/widgets.dart';
-
 import 'package:flutter/material.dart';
-
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-
 import 'package:riverpod_annotation/riverpod_annotation.dart'
     show ProviderListenable, ProviderOrFamily;
-
+import 'package:autoverpod/autoverpod.dart';
 import 'package:autoverpod_example/user_profile.dart';
 
 // ============================================================================
@@ -201,9 +190,8 @@ class AsyncUserProfileProxyWidgetRef {
 
   Selected select<Selected>(Selected Function(UserProfileState) selector) =>
       _ref.watch(
-        asyncUserProfileProvider(
-          _params,
-        ).select((value) => selector(value.requireValue)),
+        asyncUserProfileProvider(_params)
+            .select((value) => selector(value.requireValue)),
       );
 
   WidgetRef get widgetRef => _ref;
@@ -1061,9 +1049,8 @@ class UserProfileSelect<Selected> extends ConsumerWidget {
     final idValue = id ?? scope?.id;
     assert(idValue != null, 'No id provided for UserProfileProvider');
     final params = idValue!;
-    final selectedProvider = userProfileProvider(
-      params,
-    ).select((value) => selector(value));
+    final selectedProvider = userProfileProvider(params)
+        .select((value) => selector(value));
     if (onStateChanged != null) {
       ref.listen(selectedProvider, (prev, next) {
         if (prev != next) onStateChanged!(prev, next);